# backend/documentation/openapi.yaml
openapi: 3.0.0
info:
  title: Inventory Management API
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:8000
    variables: {}
paths:
...
  /warehouse:
    get:
      tags:
        - Warehouse
      summary: Read Warehouse
      description: Read all active warehouses
      operationId: getWarehouse
      parameters: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/getAll1'
                  - example:
                      warehouses:
                        - warehouse_id: 1
                          warehouse_name: Main Warehouse
                          location: City A
                          warehouse_category: Main
                          product_category:
                            - Clothing
                            - Bedding
                          created_at: '2023-11-17T12:00:00'
                          updated_at: '2023-11-17T12:00:00'
                        - warehouse_id: 2
                          warehouse_name: Baby Warehouse
                          location: City B
                          warehouse_category: Baby
                          product_category:
                            - Baby Clothing
                            - Baby Bedding
                          created_at: '2023-11-17T12:30:00'
                          updated_at: '2023-11-17T12:30:00'
                        - warehouse_id: 3
                          warehouse_name: Home Warehouse
                          location: City C
                          warehouse_category: Home
                          product_category:
                            - Bedding
                          created_at: '2023-11-17T13:00:00'
                          updated_at: '2023-11-17T13:00:00'
                        - warehouse_id: 4
                          warehouse_name: Nursery Warehouse
                          location: City D
                          warehouse_category: Nursery
                          product_category:
                            - Baby Bedding
                          created_at: '2023-11-17T13:30:00'
                          updated_at: '2023-11-17T13:30:00'
                        - warehouse_id: 5
                          warehouse_name: Outdoor Warehouse
                          location: City E
                          warehouse_category: Outdoor
                          product_category:
                            - Baby Gear
                          created_at: '2023-11-17T14:00:00'
                          updated_at: '2023-11-17T14:00:00'
                        - warehouse_id: 6
                          warehouse_name: Feeding Warehouse
                          location: City F
                          warehouse_category: Feeding
                          product_category:
                            - Feeding
                          created_at: '2023-11-17T14:30:00'
                          updated_at: '2023-11-17T14:30:00'
                        - warehouse_id: 7
                          warehouse_name: Safety Warehouse
                          location: City G
                          warehouse_category: Safety
                          product_category:
                            - Safety
                          created_at: '2023-11-17T15:00:00'
                          updated_at: '2023-11-17T15:00:00'
                        - warehouse_id: 8
                          warehouse_name: Feeding Warehouse
                          location: City H
                          warehouse_category: Feeding
                          product_category:
                            - Feeding
                          created_at: '2023-11-17T15:30:00'
                          updated_at: '2023-11-17T15:30:00'
                        - warehouse_id: 9
                          warehouse_name: Feeding Warehouse
                          location: City I
                          warehouse_category: Feeding
                          product_category:
                            - Feeding
                          created_at: '2023-11-17T16:00:00'
                          updated_at: '2023-11-17T16:00:00'
                        - warehouse_id: 10
                          warehouse_name: Maternity Warehouse
                          location: City J
                          warehouse_category: Maternity
                          product_category:
                            - Maternity
                          created_at: '2023-11-17T16:30:00'
                          updated_at: '2023-11-17T16:30:00'
              example:
                warehouses:
                  - warehouse_id: 1
                    warehouse_name: Main Warehouse
                    location: City A
                    warehouse_category: Main
                    product_category:
                      - Clothing
                      - Bedding
                    created_at: '2023-11-17T12:00:00'
                    updated_at: '2023-11-17T12:00:00'
                  - warehouse_id: 2
                    warehouse_name: Baby Warehouse
                    location: City B
                    warehouse_category: Baby
                    product_category:
                      - Baby Clothing
                      - Baby Bedding
                    created_at: '2023-11-17T12:30:00'
                    updated_at: '2023-11-17T12:30:00'
                  - warehouse_id: 3
                    warehouse_name: Home Warehouse
                    location: City C
                    warehouse_category: Home
                    product_category:
                      - Bedding
                    created_at: '2023-11-17T13:00:00'
                    updated_at: '2023-11-17T13:00:00'
                  - warehouse_id: 4
                    warehouse_name: Nursery Warehouse
                    location: City D
                    warehouse_category: Nursery
                    product_category:
                      - Baby Bedding
                    created_at: '2023-11-17T13:30:00'
                    updated_at: '2023-11-17T13:30:00'
                  - warehouse_id: 5
                    warehouse_name: Outdoor Warehouse
                    location: City E
                    warehouse_category: Outdoor
                    product_category:
                      - Baby Gear
                    created_at: '2023-11-17T14:00:00'
                    updated_at: '2023-11-17T14:00:00'
                  - warehouse_id: 6
                    warehouse_name: Feeding Warehouse
                    location: City F
                    warehouse_category: Feeding
                    product_category:
                      - Feeding
                    created_at: '2023-11-17T14:30:00'
                    updated_at: '2023-11-17T14:30:00'
                  - warehouse_id: 7
                    warehouse_name: Safety Warehouse
                    location: City G
                    warehouse_category: Safety
                    product_category:
                      - Safety
                    created_at: '2023-11-17T15:00:00'
                    updated_at: '2023-11-17T15:00:00'
                  - warehouse_id: 8
                    warehouse_name: Feeding Warehouse
                    location: City H
                    warehouse_category: Feeding
                    product_category:
                      - Feeding
                    created_at: '2023-11-17T15:30:00'
                    updated_at: '2023-11-17T15:30:00'
                  - warehouse_id: 9
                    warehouse_name: Feeding Warehouse
                    location: City I
                    warehouse_category: Feeding
                    product_category:
                      - Feeding
                    created_at: '2023-11-17T16:00:00'
                    updated_at: '2023-11-17T16:00:00'
                  - warehouse_id: 10
                    warehouse_name: Maternity Warehouse
                    location: City J
                    warehouse_category: Maternity
                    product_category:
                      - Maternity
                    created_at: '2023-11-17T16:30:00'
                    updated_at: '2023-11-17T16:30:00'
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/failed1'
                  - example:
                      Message: Error can't find any warehouse
              example:
                Message: Error can't find any warehouse
      deprecated: false
    post:
      tags:
        - Warehouse
      summary: Create Warehouse
      description: Add active warehouses to database
      operationId: postWarehouse
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/warehouserequest'
                - example:
                    warehouse_id: 101
                    warehouse_name: Main Warehouse
                    location: City A
                    warehouse_category: Main
                    product_category:
                      - Clothing
                      - Bedding
                    created_at: '2023-11-17T12:00:00'
                    updated_at: '2023-11-17T12:00:00'
            example:
              warehouse_id: 101
              warehouse_name: Main Warehouse
              location: City A
              warehouse_category: Main
              product_category:
                - Clothing
                - Bedding
              created_at: '2023-11-17T12:00:00'
              updated_at: '2023-11-17T12:00:00'
        required: true
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/success3'
                  - example:
                      Message: Create Warehouse Successed!
                      warehouse_id: 101
                      warehouse_name: Main Warehouse
                      location: City A
                      warehouse_category: Main
                      product_category:
                        - Clothing
                        - Bedding
                      created_at: '2023-11-17T12:00:00'
                      updated_at: '2023-11-17T12:00:00'
              example:
                Message: Create Warehouse Successed!
                warehouse_id: 101
                warehouse_name: Main Warehouse
                location: City A
                warehouse_category: Main
                product_category:
                  - Clothing
                  - Bedding
                created_at: '2023-11-17T12:00:00'
                updated_at: '2023-11-17T12:00:00'
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/failed1'
                  - example:
                      Message: Error Create Warehouse!
              example:
                Message: Error Create Warehouse!
      deprecated: false
    put:
      tags:
        - Warehouse
      summary: Update Warehouse
      description: Update a specific (by ID) active warehouse
      operationId: putWarehouse
      parameters:
        - name: id
          in: query
          description: ''
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/werehouserequest'
                - example:
                    warehouse_id: 1
                    warehouse_name: Main Warehouse
                    location: City AAA
                    warehouse_category: Main
                    product_category:
                      - Clothing
                      - Bedding
                    created_at: '2023-11-17T13:00:00'
                    updated_at: '2023-11-17T13:00:00'
            example:
              warehouse_id: 1
              warehouse_name: Main Warehouse
              location: City AAA
              warehouse_category: Main
              product_category:
                - Clothing
                - Bedding
              created_at: '2023-11-17T13:00:00'
              updated_at: '2023-11-17T13:00:00'
        required: true
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/success3'
                  - example:
                      Message: Update Warehouse Successed!
                      warehouse_id: 1
                      warehouse_name: Main Warehouse
                      location: City AAA
                      warehouse_category: Main
                      product_category:
                        - Clothing
                        - Bedding
                      created_at: '2023-11-17T13:00:00'
                      updated_at: '2023-11-17T13:00:00'
              example:
                Message: Update Warehouse Successed!
                warehouse_id: 1
                warehouse_name: Main Warehouse
                location: City AAA
                warehouse_category: Main
                product_category:
                  - Clothing
                  - Bedding
                created_at: '2023-11-17T13:00:00'
                updated_at: '2023-11-17T13:00:00'
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/failed1'
                  - example:
                      Message: Error update warehouse!
              example:
                Message: Error update warehouse!
      deprecated: false
    delete:
      tags:
        - Warehouse
      summary: Delete Warehouse
      description: Delete a specific warehouse
      operationId: deleteWarehouse
      parameters:
        - name: id
          in: query
          description: ''
          required: true
          style: form
          explode: true
          schema:
            type: integer
            format: int32
            example: 1
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/success2'
                  - example:
                      Message: Delete Warehouse Successed!
              example:
                Message: Delete Warehouse Successed!
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/failed1'
                  - example:
                      Message: Error Delete, cant find a warehouse
              example:
                Message: Error Delete, cant find a warehouse
      deprecated: false
...
components:
  schemas:
...
    failed1:
      title: failed1
      required:
        - Message
      type: object
      properties:
        Message:
          type: string
      example:
        Message: Error cant find a product
...
    success2:
      title: success2
      required:
        - Message
      type: object
      properties:
        Message:
          type: string
      example:
        Message: Delete Product Successed!
    getAll1:
      title: getAll1
      required:
        - warehouses
      type: object
      properties:
        warehouses:
          type: array
          items:
            $ref: '#/components/schemas/Warehouse6'
          description: ''
      example:
        warehouses:
          - warehouse_id: 1
            warehouse_name: Main Warehouse
            location: City A
            warehouse_category: Main
            product_category:
              - Clothing
              - Bedding
            created_at: '2023-11-17T12:00:00'
            updated_at: '2023-11-17T12:00:00'
          - warehouse_id: 2
            warehouse_name: Baby Warehouse
            location: City B
            warehouse_category: Baby
            product_category:
              - Baby Clothing
              - Baby Bedding
            created_at: '2023-11-17T12:30:00'
            updated_at: '2023-11-17T12:30:00'
          - warehouse_id: 3
            warehouse_name: Home Warehouse
            location: City C
            warehouse_category: Home
            product_category:
              - Bedding
            created_at: '2023-11-17T13:00:00'
            updated_at: '2023-11-17T13:00:00'
          - warehouse_id: 4
            warehouse_name: Nursery Warehouse
            location: City D
            warehouse_category: Nursery
            product_category:
              - Baby Bedding
            created_at: '2023-11-17T13:30:00'
            updated_at: '2023-11-17T13:30:00'
          - warehouse_id: 5
            warehouse_name: Outdoor Warehouse
            location: City E
            warehouse_category: Outdoor
            product_category:
              - Baby Gear
            created_at: '2023-11-17T14:00:00'
            updated_at: '2023-11-17T14:00:00'
          - warehouse_id: 6
            warehouse_name: Feeding Warehouse
            location: City F
            warehouse_category: Feeding
            product_category:
              - Feeding
            created_at: '2023-11-17T14:30:00'
            updated_at: '2023-11-17T14:30:00'
          - warehouse_id: 7
            warehouse_name: Safety Warehouse
            location: City G
            warehouse_category: Safety
            product_category:
              - Safety
            created_at: '2023-11-17T15:00:00'
            updated_at: '2023-11-17T15:00:00'
          - warehouse_id: 8
            warehouse_name: Feeding Warehouse
            location: City H
            warehouse_category: Feeding
            product_category:
              - Feeding
            created_at: '2023-11-17T15:30:00'
            updated_at: '2023-11-17T15:30:00'
          - warehouse_id: 9
            warehouse_name: Feeding Warehouse
            location: City I
            warehouse_category: Feeding
            product_category:
              - Feeding
            created_at: '2023-11-17T16:00:00'
            updated_at: '2023-11-17T16:00:00'
          - warehouse_id: 10
            warehouse_name: Maternity Warehouse
            location: City J
            warehouse_category: Maternity
            product_category:
              - Maternity
            created_at: '2023-11-17T16:30:00'
            updated_at: '2023-11-17T16:30:00'
    Warehouse6:
      title: Warehouse6
      required:
        - warehouse_id
        - warehouse_name
        - location
        - warehouse_category
        - product_category
        - created_at
        - updated_at
      type: object
      properties:
        warehouse_id:
          type: integer
          format: int32
        warehouse_name:
          type: string
        location:
          type: string
        warehouse_category:
          type: string
        product_category:
          type: array
          items:
            type: string
          description: ''
        created_at:
          type: string
        updated_at:
          type: string
      example:
        warehouse_id: 1
        warehouse_name: Main Warehouse
        location: City A
        warehouse_category: Main
        product_category:
          - Clothing
          - Bedding
        created_at: '2023-11-17T12:00:00'
        updated_at: '2023-11-17T12:00:00'
...
    warehouserequest:
      title: warehouserequest
      required:
        - warehouse_id
        - warehouse_name
        - location
        - warehouse_category
        - product_category
        - created_at
        - updated_at
      type: object
      properties:
        warehouse_id:
          type: integer
          format: int32
        warehouse_name:
          type: string
        location:
          type: string
        warehouse_category:
          type: string
        product_category:
          type: array
          items:
            type: string
          description: ''
        created_at:
          type: string
        updated_at:
          type: string
      example:
        warehouse_id: 101
        warehouse_name: Main Warehouse
        location: City A
        warehouse_category: Main
        product_category:
          - Clothing
          - Bedding
        created_at: '2023-11-17T12:00:00'
        updated_at: '2023-11-17T12:00:00'
    success3:
      title: success3
      required:
        - Message
        - warehouse_id
        - warehouse_name
        - location
        - warehouse_category
        - product_category
        - created_at
        - updated_at
      type: object
      properties:
        Message:
          type: string
        warehouse_id:
          type: integer
          format: int32
        warehouse_name:
          type: string
        location:
          type: string
        warehouse_category:
          type: string
        product_category:
          type: array
          items:
            type: string
          description: ''
        created_at:
          type: string
        updated_at:
          type: string
      example:
        Message: Create Warehouse Successed!
        warehouse_id: 101
        warehouse_name: Main Warehouse
        location: City A
        warehouse_category: Main
        product_category:
          - Clothing
          - Bedding
        created_at: '2023-11-17T12:00:00'
        updated_at: '2023-11-17T12:00:00'
    werehouserequest:
      title: werehouserequest
      required:
        - warehouse_id
        - warehouse_name
        - location
        - warehouse_category
        - product_category
        - created_at
        - updated_at
      type: object
      properties:
        warehouse_id:
          type: integer
          format: int32
        warehouse_name:
          type: string
        location:
          type: string
        warehouse_category:
          type: string
        product_category:
          type: array
          items:
            type: string
          description: ''
        created_at:
          type: string
        updated_at:
          type: string
      example:
        warehouse_id: 1
        warehouse_name: Main Warehouse
        location: City AAA
        warehouse_category: Main
        product_category:
          - Clothing
          - Bedding
        created_at: '2023-11-17T13:00:00'
        updated_at: '2023-11-17T13:00:00'
...